Fix module inception
